refactor(gui): split config/types.rs into per-settings-area submodules
`primer-gui/src/config/types.rs` had grown to 539 lines holding every
persisted GUI setting shape in one flat list — the `GuiConfig` umbrella,
the inference-backend block with its three API-key secret types, the
subsystem/embedder blocks, the speech block, and six small
single-purpose section structs.

Behaviour-preserving. `types.rs` becomes a directory module grouped by
settings area — the same grouping the settings modal renders as
collapsible sections:

- `types/mod.rs` (44) — the `GuiConfig` umbrella plus a flat
  `pub use <submodule>::*;` façade.
- `types/backend.rs` (198) — `BackendConfig` + `ApiKeySource` /
  `ApiKeySourceView` / `ApiKeyUpdate`.
- `types/speech.rs` (146) — `SpeechBackend` / `SttBackend` /
  `TtsBackend` / `SpeechSettings` / `SpeechLocaleOverride`.
- `types/sections.rs` (107) — the six small single-purpose blocks:
  `LearnerConfig`, `VocabConfig`, `BreakConfig`, `PersistenceConfig`,
  `UiConfig`, `DiagnosticsConfig`.
- `types/subsystems.rs` (105) — `SubsystemConfig` + `EmbedderConfig`,
  including the cfg-gated `default_embedder_kind()` (still private —
  nothing outside this module references it).

No call site changes. `config/mod.rs` already re-exports with
`pub use types::*;`, and `types/mod.rs` now globs its own submodules, so
every `crate::config::Foo` path — and the `use super::*` in the
1086-line `config/tests.rs` — resolves exactly as before. The
`#[serde(default)]` / `#[serde(tag = ...)]` attributes travel with their
types, so the on-disk `gui-config.json` shape is untouched.

Verification:
- `cargo test -p primer-gui` — 207 passed / 0 failed, identical to the
  pre-split baseline.
- External pub surface byte-identical (23 symbols), diffed against
  `main`.
- Order-insensitive line diff of all substantive (non-comment,
  non-import) lines old-vs-new is empty except for the re-export façade
  lines deliberately added.
- `cargo check -p primer-gui --no-default-features` green — exercises
  the `#[cfg(not(feature = "embedding"))]` arm of
  `default_embedder_kind()` that the default build configures out.
- `cargo clippy --workspace --all-targets -- -D warnings` clean;
  `cargo fmt --all -- --check` clean; `cargo test --workspace` green
  (51 × `test result: ok`, 0 failed).
- `cargo doc -p primer-gui --document-private-items` reports the same 18
  pre-existing intra-doc-link errors as before, none of them in
  `config/types/` — the new module docs add nothing.

Docs updated: CLAUDE.md's hybrid-retrieval bullet now points at
`config/types/subsystems.rs` for `default_embedder_kind()`, the
`primer-gui` `config`-module note records the nested split, and
docs/devel/android-test-apk-build.md's Step 2 path is corrected. The
frozen `docs/superpowers/plans/` and `docs/handoffs/` copies are left
as-is by convention. README/ROADMAP need no change — this alters no
capability and no user-facing behaviour.
